Create README.md
I found a book called " The Medical Consequences of Loneliness " with an empty envelope from Ruth. And a return address of Ottawa Canada. Ruth was unforgettable and we had met 40+ years ago ( I am 83 now)  .  I know nothing about GIT Hub but it was the only place I found a photo. Hope this is the Ruth I am looking for.